allTilesConfig in VideoChatSection
